Organize components and add landing page pricing
