+++ src/components/Navbar/Navbar.jsx
@@ -1,5 +1,5 @@
 import './Navbar.css';
-import { navLinks } from '../data';
+import { navLinks } from '../../data';
 
 const Navbar = () => {
   return (
